backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
openai==1.54.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
